+++ roff.c
@@ -600,7 +600,7 @@ roff_parse(struct roff *r, const char *buf, int *pos)
 	t = (r->current_string = roff_getstrn(r, mac, maclen))
 	    ? ROFF_USERDEF : roff_hash_find(mac, maclen);
 
-	*pos += maclen;
+	*pos += (int)maclen;
 	while (buf[*pos] && ' ' == buf[*pos])
 		(*pos)++;
 
@@ -1137,7 +1137,7 @@ roff_rm(ROFF_ARGS)
 
 	cp = *bufp + pos;
 	while ('\0' != *cp) {
-		name = roff_getname(r, &cp, ln, cp - *bufp);
+		name = roff_getname(r, &cp, ln, (int)(cp - *bufp));
 		if ('\0' != *name)
 			roff_setstr(r, name, NULL, 0);
 	}
@@ -1372,7 +1372,7 @@ roff_setstr(struct roff *r, const char *name, const char *string,
 	 * One additional byte for the '\n' in multiline mode,
 	 * and one for the terminating '\0'.
 	 */
-	newch = strlen(string) + (multiline ? 2 : 1);
+	newch = strlen(string) + (multiline ? 2u : 1u);
 	if (NULL == n->string) {
 		n->string = mandoc_malloc(newch);
 		*n->string = '\0';
@@ -1383,7 +1383,7 @@ roff_setstr(struct roff *r, const char *name, const char *string,
 	}
 
 	/* Skip existing content in the destination buffer. */
-	c = n->string + oldch;
+	c = n->string + (int)oldch;
 
 	/* Append new content to the destination buffer. */
 	while (*string) {
